+++ system/helpers/form_helper.php
@@ -445,7 +445,13 @@ if ( ! function_exists('form_submit'))
 	 */
 	function form_submit($data = '', $value = '', $extra = '')
 	{
-		$defaults = array('type' => 'submit', 'name' => ( ! is_array($data) ? $data : ''), 'value' => $value);
+		$defaults = array('type' => 'submit', 'value' => $value);
+
+		if ( ! is_array($data) && ! empty($data)) 
+		{
+			$data = array('name'=>$data);
+		}
+
 		return '<input '._parse_form_attributes($data, $defaults).$extra." />\n";
 	}
 }
@@ -483,7 +489,13 @@ if ( ! function_exists('form_button'))
 	 */
 	function form_button($data = '', $content = '', $extra = '')
 	{
-		$defaults = array('name' => ( ! is_array($data) ? $data : ''), 'type' => 'button');
+		$defaults = array('type' => 'button');
+
+		if ( ! is_array($data) && ! empty($data)) 
+		{
+			$data = array('name'=>$data);
+		}
+		
 		if (is_array($data) && isset($data['content']))
 		{
 			$content = $data['content'];